Refactor to a public record manager interface

+++ SongPlayHistory/SongPlayData/RecordsManager.cs
@@ -13,9 +13,9 @@ using SongPlayHistory.Model;
 using SongPlayHistory.Utils;
 using Zenject;
 
-namespace SongPlayHistory
+namespace SongPlayHistory.SongPlayData
 {
-    internal class RecordsManager: IInitializable, IDisposable
+    internal class RecordsManager: IInitializable, IDisposable, IRecordManager
     {
         private readonly string DataFile = Path.Combine(UnityGame.UserDataPath, "SongPlayData.json");
 
@@ -32,7 +32,7 @@ namespace SongPlayHistory
                 return;
             }
             
-            if (!LoadRecords(DataFile) || Records.Count == 0)
+            if (!LoadRecords(DataFile, out var records) || records.Count == 0)
             {
                 _logger.Warn("Did not load any records from file. Will try to restore from a backup.");
                 // Try to restore from a backup.
@@ -40,7 +40,8 @@ namespace SongPlayHistory
                 if (backup.Exists && backup.Length > 0)
                 {
                     _logger.Notice("Restoring from a backup");
-                    LoadRecords(backup.FullName);
+                    LoadRecords(backup.FullName, out records);
+                    Records = records;
                 }
                 else
                 {
@@ -49,9 +50,13 @@ namespace SongPlayHistory
                     Records = new Dictionary<LevelMapKey, IList<Record>>();
                 }
             }
+            else
+            {
+                Records = records;
+            }
             
             SaveRecordsToFile();
-            _logger.Info($"Loaded {Records.Select(pair => pair.Value.Count).Sum()} records from {Records.Count} levels.");
+            _logger.Info($"Loaded {SumRecords(Records)} records from {Records.Count} levels.");
             
             // TODO remove bad records?
 
@@ -61,21 +66,22 @@ namespace SongPlayHistory
             BSEvents.LevelFinished += OnLevelFinished;
         }
 
-        private bool LoadRecords(string path)
+        private bool LoadRecords(string path, out Dictionary<LevelMapKey, IList<Record>> records)
         {
             _logger.Debug($"Loading history from {path}");
             try
             {
                 // Read records from a data file.
                 var text = File.ReadAllText(path);
-                var records = JsonConvert.DeserializeObject<Dictionary<LevelMapKey, IList<Record>>>(text, new RecordJsonConvertor());
-                Records = records ?? throw new Exception();
+                var deserialized = JsonConvert.DeserializeObject<Dictionary<LevelMapKey, IList<Record>>>(text, new RecordJsonConvertor());
+                records = deserialized ?? throw new Exception();
                 return true;
             }
             catch (Exception e)
             {
                 _logger.Error("Unable to deserialize song play records.");
                 _logger.Error(e);
+                records = new Dictionary<LevelMapKey, IList<Record>>();
                 return false;
             }
         }
@@ -135,23 +141,28 @@ namespace SongPlayHistory
             }
         }
 
-        public IList<Record> GetRecords(IDifficultyBeatmap beatmap)
+        public IList<ISongPlayRecord> GetRecords(IDifficultyBeatmap beatmap)
         {
-            var config = PluginConfig.Instance;
-
-            if (Records.TryGetValue(new LevelMapKey(beatmap), out var records))
+            var key = new LevelMapKey(beatmap);
+            return GetRecords(key);
+        }
+        
+        public IList<ISongPlayRecord> GetRecords(LevelMapKey key)
+        {
+            _logger.Debug($"Getting records for {key}");
+            if (Records.TryGetValue(key, out var records))
             {
-                // LastNote = -1 (cleared), 0 (undefined), n (failed)
-                var processed =
-                    from record in records
-                    where config.ShowFailed || record.LastNote <= 0
-                    orderby config.SortByDate ? record.Date : record.ModifiedScore descending
-                    select record;
-                
-                return processed.ToList();
+                _logger.Debug($"Total number of records: {records.Count}");
+                return records.Copy();
             }
 
-            return new List<Record>();
+            _logger.Debug("No records found.");
+            return new List<ISongPlayRecord>();
+        }
+        
+        public IDictionary<LevelMapKey, IList<ISongPlayRecord>> GetAllRecords()
+        {
+            return Records.ToDictionary(pair => pair.Key, pair => pair.Value.Copy());
         }
 
         private void SaveRecord(IDifficultyBeatmap? beatmap, LevelCompletionResults? result, bool isMultiplayer, bool energyDidReach0, ScoreRecord? failRecord)
@@ -182,8 +193,8 @@ namespace SongPlayHistory
             _logger.Debug($"Cleared: {cleared}, NoFail: {noFailEnabled}, SoftFailed: {energyDidReach0}, FailRecord: {failRecord}");
             
             var param = ParamHelper.ModsToParam(result.gameplayModifiers, energyDidReach0);
-            param |= submissionDisabled ? Param.SubmissionDisabled : 0;
-            param |= isMultiplayer ? Param.Multiplayer : 0;
+            param |= submissionDisabled ? SongPlayParam.SubmissionDisabled : 0;
+            param |= isMultiplayer ? SongPlayParam.Multiplayer : 0;
             var time = DateTimeOffset.Now.ToUnixTimeMilliseconds();
             Record record;
             
@@ -196,20 +207,21 @@ namespace SongPlayHistory
                     ModifiedScore = failRecord.Value.RawScore,
                     RawScore = failRecord.Value.ModifiedScore,
                     LastNote = failRecord.Value.NotesPassed,
-                    Param = (int) param,
+                    Params = param,
                     MaxRawScore = failRecord.Value.MaxRawScore
                 };
             }
             else if (!cleared && noFailEnabled && energyDidReach0 && failRecord != null)
             {
-                // No fail is enabled, but level still failed (for example, thr FailButton mod)
+                // No fail is enabled and did soft fail, but level still failed (for example, the FailButton mod)
+                // ues our tracked values at the time of soft fail
                 record = new Record
                 {
                     Date = time,
                     ModifiedScore = failRecord.Value.RawScore,
                     RawScore = failRecord.Value.ModifiedScore,
                     LastNote = failRecord.Value.NotesPassed,
-                    Param = (int) param,
+                    Params = param,
                     MaxRawScore = failRecord.Value.MaxRawScore
                 };
             }
@@ -221,12 +233,12 @@ namespace SongPlayHistory
                     ModifiedScore = result.modifiedScore,
                     RawScore = result.multipliedScore,
                     LastNote = cleared ? -1 : result.goodCutsCount + result.badCutsCount + result.missedCount,
-                    Param = (int) param,
+                    Params = param,
                     MaxRawScore = cleared ? null : failRecord?.MaxRawScore
                 };
             }
 
-            _logger.Info($"Saving result. Record: {record.ToShortString()}");
+            _logger.Info($"Saving result. Record: {record}");
 
             var key = new LevelMapKey(beatmap);
 
@@ -269,10 +281,10 @@ namespace SongPlayHistory
             var backupFile = Path.ChangeExtension(DataFile, ".bak");
             try
             {
-                if (File.Exists(backupFile))
+                if (File.Exists(backupFile) && LoadRecords(backupFile, out var backupRecords))
                 {
                     // Compare file sizes instead of the last modified.
-                    if (new FileInfo(DataFile).Length > new FileInfo(backupFile).Length)
+                    if (SumRecords(Records) >= SumRecords(backupRecords))
                     {
                         File.Copy(DataFile, backupFile, true);
                     }
@@ -290,6 +302,11 @@ namespace SongPlayHistory
             {
                 _logger.Error(ex.ToString());
             }
+        }
+
+        private static int SumRecords(Dictionary<LevelMapKey, IList<Record>> records)
+        {
+            return records.Select(pair => pair.Value.Count).Sum();
         }
     }
 }
